// frontend/src/app/apiClient/model/casePatientProgress.ts
/**
 * LoopBack Application
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { BigDecimal } from 'bigdecimal';







export interface VitalSignsStatus {
    userID: string;
    bloodPressure: string;
    heartRate: string;
    respiratoryRate: string;
    temperature: string;
    bloodSugar: string;
    oxygenSaturation: string;
    oxygenFlowRate: string;
    measurementMode: string;
    painMeasurement: BigDecimal;
    date: string;
}


export interface DeliveredMedications {
    userID: string;
    description: string;
    dose: string;
    route: string;
    patientResponse: string;
    date: string;
}


export interface ProgressNote {
    userID: string;
    text: string;
    date: string;
}






export interface CasePatientProgress { 
    patientProgressID: string;
    caseID: string;

    escort1ID?: string;
    escort1Signature?: string;
    escort2ID?: string;
    escort2Signature?: string;
    medicalProviderName?: string;
    medicalProviderSignature?: string;
    medicalProviderSignatureDate?: string;

    patientBelongings: boolean;
    patientBelongingsDesc?: string;

    statusUpdates?: Array<VitalSignsStatus>;
    medications?: Array<DeliveredMedications>;
    notes?: Array<ProgressNote>;
}